fixed a bug; added feature to show coded media also in code_text
